House cleaning.

+++ ethers/src/Transaction.m
@@ -403,8 +403,8 @@ static NSData *NullData = nil;
 #pragma mark - NSObject
 
 - (NSString*)description {
-    return [NSString stringWithFormat:@"<Transaction to=%@ from=%@ nonce=%ld gasPrice=%@ gasLimit=%@ value=%@ data=%@ chainId=%d signature=%@>",
-            self.toAddress, self.fromAddress, (unsigned long)self.nonce, [self.gasPrice decimalString], [self.gasLimit decimalString],
+    return [NSString stringWithFormat:@"<Transaction to=%@ from=%@ nonce=%d gasPrice=%@ gasLimit=%@ value=%@ data=%@ chainId=%d signature=%@>",
+            self.toAddress, self.fromAddress, (int)self.nonce, [self.gasPrice decimalString], [self.gasLimit decimalString],
             [self.value decimalString], [SecureData dataToHexString:self.data], _chainId, _signature];
 }
